Chapter 13: Creating User Profile Page

+++ server/controllers/user.controller.js
@@ -20,6 +20,12 @@ export const getUser = async (req, res, next) => {
 };
 
 export const updateUser = async (req, res, next) => {
+  if (req.user.id !== req.params.id) {
+    return next({
+      status: 401,
+      message: 'You can only update your own account',
+    });
+  }
   try {
     if (req.body.password) {
       req.body.password = await bcrypt.hash(req.body.password, 10);
@@ -44,10 +50,16 @@ export const updateUser = async (req, res, next) => {
 };
 
 export const deleteUser = async (req, res, next) => {
+  if (req.user.id !== req.params.id) {
+    return next({
+      status: 401,
+      message: 'You can only delete your own account',
+    });
+  }
   try {
     const query = { _id: new ObjectId(req.params.id) };
     await collection.deleteOne(query);
-    res.status(200).json({ message: 'User has been deleted!' });
+    res.status(200).json({ message: 'User has been deleted' });
   } catch (error) {
     next({ status: 500, error });
   }
